Move some play state over to MpvWidget
is_playing and is_paused continue to be tracked by the ui code, as it is
a sort of tri-state thing at present.  And as for speed, mpv should
never be changing that unless something is being mucked around with in a
script, which should not be an issue.

I'll address some player logic in the next commit.