Center the screen more correctly.
In a multi-monitor case where the left screen was switched off and the
right screen is the primary monitor, scaled resizing would always place
the window at the left most edge.  So instead of emitting resize and
movement function calls, set the geometry directly and attempt to place
it on the same screen as the mouse cursor on the first run.  This
however requires some hackery in the constructor.