hostwindow: connect chapter navigation buttons and rename their actions
mpc-hc's menu is confusing, these are clearly chapter forward/backward
menu items but not under the chapter menu like you would suppose, but
are top-level items under the navigate menu.  It's quite easy to assume
that signal handlers for chapter navigation do not exist because they
don't (navigate_chapter_next is actually navigate_next).  Which is why
this wasn't in the last commit.

Now as for which notation you prefer, following the menu layout
convention as-is or renaming to the logical layout while breaking the
menu layout convention, is up to you.  In the end I've chosen
maintainability of the source code over the menu layout convention.