hostwindow: convert to_date_fmt into a member function
With the new sorting of functions, this helper function needs to be
placed after everything and have a forward-declaration, rather than
(arguably incorrectly) defining it just before we need it.

Also starring: rename globalize_actions to action_globalize_all, to fit
in more with the type_name_event convention we've got going on.  Movie
directed by Pe Dan Tic.