drawnslider: add to project
Reimplement QMediaSlider as a class based on QDrawnSlider, an abstract
class that which manages code and geometry calculations common to custom
sliders.  This commit also includes QVolumeSlider as well, which was the
main purpose of performing this refactor.

This is actually quite similar in logic and functionality to the
previous way, but better in terms of abstraction and extendability.  It
might not be as straightforward to newcomers as adding a plain
QVolumeSlider module with copy-pasted code, but who cares; nobody reads
my repo anyway.