drawnslider: set minimum height
The main window's layout requires this to be set, or else it will
collapse into nothingness.  Previously we were testing it on it a window
with plenty of room to spare, where providing the setMaximumHeight hint
gave it the desired appearance.  The MainWindow layout of this project
however is slightly more demanding.