mainwindow: add ability to hide borders to view>hide_menu
This skips the setting of having a bordered titleless window because
there is no way to guarantee it across every window managers.  I'd
rather not go the chrome route of drawing my own decorations just so I
can do that.

Also this action is not supposed to be checkable, it's supposed to
change.