mpvwidget: restyle interface closer to Qt style guidelines
I'll be honest, it bothered me that I was using inconsistent casing in
my clases. At least now it looks like a Qt class rather than a library
wrapper.

Also if there's one thing to come out of this it is that I remembered
that I was tracking play/pause/stop state in the ui code and not over
here.

See:
https://wiki.qt.io/Qt_Coding_Style
https://wiki.qt.io/Coding_Conventions
http://doc.qt.digia.com/qq/qq13-apis.html
http://geosoft.no/development/cppstyle.html