mainwindow: prefer calling updateSize from setSizeFactor, not slots
Instead of the view also being the controller, (which is extremely
attractive for small scale things,) I'm now trying to convert mainwindow
into a generic widget and use an actual controller class to manage
playback.  This will require spinning off code from here into a
PlaybackManager class.

The ideal I'm aiming for is for each mainwindow slot to call one
function of the manager per event.  Therefore the mainwindow would
respond to changes in the manager's state and less in relation to
itself, at least partially.

The non idealness of the current situation becomes obvious when you
realize that the exposure of some private data members as properties
seem pointless, and would not be very useful as external properties at
all, or are simply used to track internal state.  This implies that an
object is missing; in this case, the manager is missing.

Coincidentally, including a manager class will make integrating
playlists easier.  See the codebase.svg in the DOCS directory for an
idea of what the program with this would look like.