playlist: add to project
A set of classes that behave like playlists.  What notable about this
module is the liberal use of uuid's associated with each item and
playlist.  This is so that we do not fail if the currently playing item
has been removed from all playlists, and makes displaying a list widget
of this easier.

Perhaps needs some polish to eliminate out-of-class direct member
access, but I'll get to that.