manager: remove knowledge of mainwindow requirement from manager
After some guru meditation leading to the conclusion that mainwindow
should not know anything about what it is communicating with, relabel
most public functions of the manager so that it can be connected with
signals emitted from mainwindow.

These signals and slots will need to be connected by the instantiator,
namely, the flow object in main.

See http://doc.qt.io/qt-4.8/signalsandslots.html for more info.