mainwindow: emit signal on stop instead of commanding mpvw directly
The idea we're running with (in case you haven't noticed), is that the
mainwindow will merely emit changes in the ui state and/or ui commands
made, and rely on signals back from the manager to reflect any property
changes.  For example, this is why now frameskipping actually shows the
window unpausing for a brief period of time while the player advances to
the next frame, whereas it used to be fudged into a paused state while
frameskipping.

The benefit of this approach is that mainwindow can esentially stay as a
player wrapper while the playback logic can be tweaked elsewhere as much
as we like.

There isn't much to go after this and playlist management, other than
persistent storage (preferences and recent files) and subtitle
management.  Hopefully by then mpv will get vo properties to make the
inevitable preferences dialog a bit more straightforward.  There's also
the favorites menu and disc/streaming management, but who cares about
that.