playlistwindow ui: hash out the playlist user interface
So far there have been four layouts for this I have entertained:
- A single playlist widget with a user-managed dropdown switcher
- Stacked tabs with a seperate list to select the displayed playlist
- Paged with a toolbox widget
- The current tabbed one
We're going with the tabbed behaviour because that seems less unobvious.
I might change it to paged though if program use shows that tabbing is
not as optimum.

I need icons to make the buttons unobtrusive, so the window will be
wider than it should be until I get them done.

Also, synchronizing commits between a laptop and a dead computer is a
real bitch.