mainwindow: workaround main thread freeze and remember last directory
Spawn a File Open dialog and connect to it asynchronally.  Invoking it
by its static function was intermittentally freezing the thread, and the
program would need to be manually killed.  This splits the function in
two and places the result parsing in another slot.

Also open at the directory of the last open file, which is the actual
itch I was scratching this coding session.