mainwindow: simplify (quick) file open
Spin off common code to a helper function, and add a new slot to signify
that that a file was opened quickly.

This removes the need for the old function that handled the file
dialog's selection signal, which could not discern the difference
between a quick open and a regular open.  Note with this way of using
lambdas in connect statements, we can get rid of DataEmitter.