manager,mpvwidget: fixup opening file over paused running file
If a file is paused while a file is opened, it will not start playing
unless mpv's internal state is unpaused.  Unpause mpv after each file
load so that it is guaranteed to run, which is probably what the user
wants when loading a file.

Also remove resetting the tracked mpv playback speed, as that seems to
be preserved by mpv across file runs.