playlist: make string(list) converters m3u-compat parsers
Because we are using a QVariant structure to preserve playlist
information, don't attempt to preserve all information with the to/from
string(list) functions.  Intead only take the url as provided in a text
file, and generate the other stuff at runtime.

Also it seems not everything relating to the last commit got through.
Oh well.  I don't care so much to rewrite history just to get rid of
that.