playlistwindow,qdrawnplaylist: clear the quick playlist on removal
The quick playlist can never be removed.  When the user the clicks
remove playlist, remove all of its items instead of ignoring it.

Now that it's necessary to call qdrawnplaylist's remove all code outside
of qdrawnplaylist, expose the remove all lamda code as a common
function.

Why was it a lamda and not an function in the first place?  You may note
that I prefer keeping program logic compartmentalized rather than
introducing function and module sprawl.  IMHO, there's litte worse than
downloading a project and wondering what on earth is going on.