helpers: wait for acknowledgement from server_newConnection
In the strange corner-case where the server can be connected to, but is
unresponsive, (likely due to a crash,) previous behavior would result in
a quick termination with nothing getting done.  Instead wait for ACK and
only then terminate.

On the other hand, if reading takes too long, a second instance will
genuinely be spawned.  But if the main thread of the previous instance
is frozen, that might be a good thing anyway.