mpvwidget,manager: change to buffering state before playback is started
Differentiate between file loading and actual playback, such that it
obvious something is happening when a stream is loading but not playing,
so tell the ui that it is buffering while we're waiting for a file to
load.  This is not entirely accurate however, because network buffering
is different from loading a file.