main: pass through present working directory with payload
Discovered while testing parameter passing on the commandline.  Make
sure the present working directory of the new process is passed through
to the old process, so the old process knows where relative file paths
should resolve from.  Previously we would use the present working
directory of the existing process to parse parameters, which was not
always be the same.