mpvwidget: move most mpv interaction to a seperate thread
This keeps the main thread clean, as property changes tend to appear out
of the ether rather than taking time on the main thread.  Fixes the root
cause behind my filing mpv #2529.

As necessity of porting this, most of the code had to be made more
generic, because half of this code was written earlier this week.  This
code had QVariant semantics rather than the mpv_format and pointer
semantics.  This port may have introduced some bugs, which need to
squashed.

On the bright side, the event parsing routines no longer looks like
spaghetti.  Look at that neatness.