mpvwidget: schedule update after video playback ends
Don't depend on mpv to update ourselves after playback is finished.
Sometimes we receive the last update request after we tell ourselves to
logo draw, some times afterward.  It would only display correctly in the
latter case.  Therefore call update() when we think we should be drawing
a logo instead.

This is caused by the latency between the time we receive events and
their actually getting processed by the main thread. (this is a good
thing, because it prevents the gui from getting blocked.)