mpvwidget: remember play time and play length
These have to be tracked internally because
- they are needed while updating the statusbar
- duration and position change seperately
- getting properties is expensive

The real reason is that I forgot to notice exactly what my code was
doing for the first two property changes in the old code.  Or more
correctly, I did notice it.  "Wow that's strange, why am I storing that
only to emit it later?  I'm not doing that for the other ones.  I know
what to do, I'll eliminate that when I rewrite it."  Sigh.  I guess the
new code IS a lot simpler, but these two still needs to be tracked
instead of emitted straight away.