settingswindow: be more tolerant of dud json configs
Fall back to sensible defaults when getting an out-of-bounds settings
value.  Instead of primitive C types and (array of) array of char
pointers and boundless indexed lookups, use (QLists of ) QStringLists
and C++11 initializer lists and crafted queries that accept a default
value, respectively.

On another note, there has to be a better way than storing defaults in
several different places.