docs: add another codebase layout illustration
I'm bothers me  that I'm using Manager to send signals to the main
window and mpvwidget.  Manager arguably has no reason to know what
the VO commandline is for instance, when Manager was meant to
handle playback and playlist navigation.  However synchronizing
settings between the multiple sources of setting changes and their
targets seems a bit like spaghetti.  The solution to this is of
course simple, but does not seem to be obvious.  Because honestly,
who wants to write a bunch of boilerplate. (which is what we seem
to be doing at the moment.)