main/manager: remove boilerplate from settings->mpvwidget connection
Prefer direct connections between components.  Ownership etc was being
too strictly enforced and producing pointless boilerplate, which led to
stagnation and general unmotivation on the part of the developer.

Using more direct connections has the added benefit of reducing the
amount of required MpvWidget knowledge in Manager.  We'll add these
functions back if they are needed, but I don't expect to quite now.