settingswindow: fix typo in placeholder string
%a and %b is used for range, not %i and %j.  At one stage we used to %i
and %j to denote beginning and end, with %a as an indication for audio
track presence.  While this may make some sense to mathematicians and
programmers, it likely would not have made any sense to the lay-man.
Therefore duration format specifiers were changed to something more
familiar, with %d standing for deleted tracks.  The placeholder string
was not updated to reflect these changes, and this commit fixes that.