settingswindow: change zoom preset from settings dialog
Changes made in the settings dialog wrt zoom presets are sent back to
the main window.

This commit also adds autofit smaller to the settings dialog, which was
missing from the gui because it was only added to the codebase in a
recent commit.  And it fixes the logic bug in mainwindow that prevented
autofit smaller from working as it should.

Fixes #2.