mainwindow: minor data duplication
Don't keep track of preset data in more than one location.  This
paradigm led me to not manually feed the preset data back from settings
window via setSizeFactor, setZoomMode and setFitFactor, which would have
required the arrays inside of this commit to be inside settingswindow.
Instead zoom presets are passed around by index, and expanded to their
particulars in one place.

Of peculiar note is that Disable Snapping preset internally has a
negative index because it is in fact not a preset at all, but simply
means hands off wrt the autozoom code.