manager,playlist*: navigate to previous
Navigate to previous file instead of the next file during chapter
navigation.  When attempting to navigate to a point that does not exist
in a file, we behave like the file had ended and skip to the next file.
Instead check for being in the zeroth chapter in the previous navigation
case, and if so skip to the previous file.

Fixes #7.