manager: always advance with navigate->next when repeating
Check for beyond-the-end navigation and if so, force a playlist
advancement.  Instead of relying on mpv to stop playback with past-the-
end chapter navigation and then parsing the playback finished case,
straight away ask for the next file during the final chapter.  The
finished case had no way of telling if it was caused by actually
reaching EOF or navigating too far, and would loop continually.  This is
no longer the case, because we never try to navigate too far
(hopefully).