helpers: add display format parser
Scans through the format string and makes substitutes as necessary.  I
suppose I could make a linked list/tree of the format string and then
use it as a generator, but let's leave that for something later.

Also moves the shared grabBrackets lamda out of the other parser into a
proper function.