helpers: add yet another display parser
Second version of the previous commit.  Parse the format specifier once,
then use the resulting structure to quickly pump out a display string.

This is about 2.3x the size but will supposedly be faster because the
output section is only 29 lines long.  So therefore we have a one-off op
of about 125 lines per "Apply" in the settings dialog and only 29 later,
rather than 105 lines per each time a playlist item is displayed.  Seems
to be a nice trade-off.  If this works I'll revert the previous commit.