playlistwindow,qdrawnplaylist: use customized playlist format
Store a local DisplayParser object in the PlaylistWindow object and pass
it to the tabs by a pointer.  The tabs use the display parse to invoke
its output function.  The playlist window may look slightly inconsistent
with tracks that do not have any metadata loaded yet.  It may be
possible to fix that by spawning so-called MpvInspector threads for
unscanned files.

There were a few typos in the previous commits that were expressed at
runtime, which this commits also fixes.

Fixes #6.