settingswindow: turn off autozooming by default for tiling wms
Included is a shortlist that tries to detect what wm is being used. May
spectacularly fail if the environment isn't sane.  Note there may be a
better solution that this, but I don't feel like tracking it down
straight away.

Eventually we may enable the more proper pan'n'scan mpv-powered zooming,
in which case the zoom should be able to be set independent of the
window size.  In the meantime I encourage wm users to use floating mode
if they really want autozooming.