storage: write settings to the right location
Instantiate the constructor in full.  Due to an accident in the calling
the constructor, settings in the Windows build were written to C:/
HKEY_CURRENT_USER/SOFTWARE/mpc-qt.  Although somewhat of a neat concept,
this is probably not what people want.

PSA: you'll lose your playlists and settings unless you copy them over
to the AppData directory. (Roaming/mpc-qt on my machine.)