helpers: add MouseState and Command to project
MouseState is used to track the mouse state + keyboard modifiers at
various times.

Command is a tuple of keyboard shortcuts and mouse states associated
with a specific action.

The idea is that a widget will intercept the mouse press/release and
wheel events, convert them to a MouseState, and fire an action if it is
mapped.