main,mainwindow,settingswindow: Provide a default mouse map
The idea we have here is that we pretend we have a default mouse map in
the form that's provided by the JSON config files, and assign that to
the default key map stored in the settings window.  Do this just before
loading the actual json settings, so we can get the right variant from
the action editor before they're overwritten.