mainwindow: accept drops over the main window too
The difference between dropping over the main window and the playlist
window is that dropping over the main window will open them straight
away.  Before this, drag-and-drop only worked while over the playlist
window, and not at all otherwise.  Useful for example when having the
main window on a projector while the playlist window is closed.
(although you should if possible leave the playlist window open on
another screen so that your audience doesn't see your mouse.)