qdrawnslider: more correctly draw the grooved area on HiDPI
There used to be an ugly gap caused by drawing the highlighted area
while with the groove pen color.  While harmless at 100%, it's noticable
at 200%.  So separate the marking and fills of the groove away from each
other so that the highlighted area can be drawn completely.  This means
that the grooves and the chapter marks are drawn later after the
background fills are drawn.