playlist*: prefer QSharedPointer over naked pointers
...and use QReadWriteLock to protect data, where reading has the
connotation of accessing members, and writing has the connotation of
adding or removing members.  Therefore reading or writing can be likened
to parsing or changing lines in a text file.

Use of QSharedPointer is necessitated because we'll be searching a
playlist from another thread to tag items according to criteria, and we
don't want the playlist to be deleted out from beneath us.