playlistwindow: add quick search field
This search triggered by this field only searches for the metadata that
has been loaded, so it may end up with not finding every expected file.
It is at present always shown, making the playlist window look cluttered
at small sizes.  I'll add a few more cosmetic fixes, such as showing the
search field with ctrl+f in a few more commits.  And in the (not too?)
distant future, metadata preloading.

Fixes #4.