mainwindow: remove ancient background hax
Remove left-over black stylesheet that was interfering with the playlist
window background.  Improves appearance on Windows to native look and
feel.

Back in the day when we were embedding an mpv window, we would fiddle
around with fancing nesting and show our logo picture beneath the video
widget.  This involved setting a stylesheet for the video area, which is
no longer needed after the switch to opengl_cb.