helpers,manager,mpvwidget,qdrawnslider: add missing include files
Some files for the standard namespace are not pulled in on some versions
of gcc/libstdc++ and/or are not pulled in some versions of qt.  So
include them directly to hopefully prevent future failed builds.

Not all of these additions are actually necessary right now, only the
one in helpers.cpp, but it doesn't hurt.

Fixes build of 1cb4618.