helpers: logodrawer: use a smooth pixmap all the time
One of the reasons why we changed it earlier was because it seemed like
it was sometimes missing the mark for sizing or some much thing.  Now
that the noVideoSize case appears to be fixed in a recent commit, it no
longer needs to be this way for made-up reasons.