manager: use idling semantics
Only attempt to do anything wrt playlist advancement etc when the player
core is idle, not when playback of a particular file is finished.  This
was the principle cause of #18, where attempting to open a file quickly
would cause a race condition where we would interpret events as if they
had came in an out-of-order fashion.

Symptoms of the race condition were: runaway playlist advancement and
metadata overwriting from (semi-)adjacent tracks.