qdrawnslider: draw slider handle with fewer draw calls
Draw the handle twice, once at 4 pixels wide and one at 2 pixels wide.
Fixes seeing between the pixels esp when over a highlighted area and
squinting.

For what it's worth, reducing the pen constructors here to just the
color and the width gives us beveled corners, which is kind of neat.