mainwindow: replace cornerThing with actual size grip
Remove it and replace it with something that works.  In mpc-hc, this is
a decorative widget that displays pointless speaker icons, but in
practice it allows us to resize the window from there when the window
has lost its frame through ctrl+0.  Additionally, cornerThing's
border:none style prevented it from showing up in tab ordering too,
leading to puzzling behaviour due to it not having any indication of it
being tabbed to.  This is roughly equivalent to replacing a pimple with
a freckle.