main: fix multiple monitors window switching
Force every monitor to start on the same screen.  I turned on my second
monitor inside Linux for the first time in ages, (usually I leave it off
for my vm,) and straight away noticed a bug where the window would jump
to the other monitor on the first autosize.