docs: add paragraphs about libmpv state exposure
The internal state extension technically solves #21.

After most of this was written, it occurred to me that a better way for
third-parties is to emulate mpv's ipc instead.   Something in the short
term is still better than nothing.  If you're a serious developer and
have a significant code base, just wait for an mpv ipc emulation layer,
or use the svp-support branch.