qdrawnplaylist: include the extratimes field
When drawing a particular item, rework the quick queue area to also
display the extra times. It is a field of the form "+%d" which only
appears when there are more than one extra time configured for playback
for a particular track.  Likewise with the quick queue.

Also add an iterator for the selected items of a list, which we'll be
using for iterating over the selected items in the next commit.