qdrawnslider: inherit from QOpenGLWidget rather than QWidget
I changed something in my BIOS recently, so the cpu is running at stock
speed.  Retesting the amount of drain, the slider raises cpu by about 1%
to 4.7% while the status bar almost doubles cpu usage to around 10%.  So
the status time needs some optimization.

Also I noticed in my other recent project, darkcropper, that drawing on
to a gl widget is much faster than a standard one.  So this is a gl
widget now.