main,ipc: add input-ipc-server emulation
A filtered version of --input-ipc-server.  The implementation of which
is subject to change. e.g. allowing loadfile to succeed in some fashion
rather than block it.

Property observance is untested.  A client will receive all properties
it asks for, provided that the userdata passed through is not zero.
(zero is reserved by mpc-qt).  If you are in the interest of passing
through a possibly-NULL memory pointer, please note that most OS will
trap access to lower memory, so setting your userData to 1 or some low
value is a possible solution.  Your socket code should however be
designed in such a way that it does not need to do that.

Discussed in #22.