settingswindow.ui: prefer user-shaders over pre/post-shaders
Remove post-shaders and pre-shaders list, and replace with a single
active shader list.  There is a movement in mpv to make the shader
interface generic and thereby move complicated logic out to userland;
thus the so-called user-shaders vo parameter.  This is arguably removing
access to the pre/post shaders options, which may be inconvenient, but
people should be rewriting their pre/post-shaders into user-shaders
anyway, so let's encourage that a little.