settingswindow: save ListWidget contents, not selection
Use a list of lookups to save the widgets by property, and "override"
the ListWidget case by constructing lists out of its contents.  And
similar with restoring.  We have to do this because there is no property
that contains ALL of the items for ListWidgets.  (I guess to discourage
heavy abuse?)

I'm not sure if this way is at all faster, but the save/restore
mechanism is simpler, and I don't mind a little bit of expense every now
and then if the notation is simpler.  Especially if the expense is
contrained to initalizers.

With this commit, the list of shaders is preserved.