mainwindow: change logic for showing and hiding the bottom area
Mouse tracking has been enabled for the playlistWindow_ and the bottomArea.

hideTimer_timeout() now has a check whether the bottomArea is currently being hovered.

checkBottomArea() now takes the global cursor position as an argument and mapping to the local position is done in this function in a context depending on the selected bottomAreaBehavior.