mpvwidget: always paint, even when minimized
Proxy calls to update() through maybeUpdate(), which simulates a draw
operation when the window is minimized, or calls the actual update()
when the window is shown.  The custom draw+swap operation effectively
simulates having an infinite framerate display.

Note that the last time I did manual frameswapping like this, Windows
would crash.  But I can't test it on Windows at the moment due to a
broken video card.

With this patch, frames are no longer dropped when minimized.  Fixes
#42.