qdrawnplaylist: properly handle the insert case
Create parentless PlayItems when inserting, and rework the PlayItem
constructor to work around verification of item validity in the
parentless case.

The parent of the list item has to be NULL when inserting at arbitary
locations, so we can't test if the item's uuid is valid when we don't
know who our parent is.  Therefore only allow setting the uuid the
displayed item once, in the constructor.  Arguably the creator should
know the desired uuid at creation time, and shouldn't be fiddling around
with it later.

Fixes #43.