mpvwidget,settingswindow,main,README: use new option interface
vo/gl related options are now regular options, and vo-cmdline has been
depreciated.  What we do now for the vo options is pass them to a
function which caches options before (optionally, on a change) sending
them to mpv.  As a result there appears to be slightly neater if then
else etc logic in the settings parser, and we don't spam libmpv anywhere
near as much.

This bumps the required mpv version to 0.21.0. (effectively git head.)