*playlist*: add quick queue widget
Introduce ItemCollection for holding all items and use it for adding
items by a single uuid to the the quick queue widget.  Also virtualize
some overriden functions.  Requires breaking the 1:1 rule from the ui
side to the data side, which I anticipated somewhat by the liberal use
of Uuids instead of pointers but not quite.  On the plus side now the
playlists don't have to own the item data, they just have to have an
idea about what's there, which shouldn't be all that hard to do.

Fixes #40.