mainwindow: implement custom version string
Adds a version field to the about box.  This field defaults to
"Development Build" when MPCQT_VERSION_STR is not set.  This definition
can be set by adding the following line to the pro file:

    DEFINES += MPCQT_VERSION_STR=\"xx.yy\"

The project tree will need a clean and rebuild for this to take effect.

This will be used for #95.