settingswindow: disable unimplemented functions
Not that these aren't impossible to do.  I merely haven't written the
code to tell mpv what they are, or the main gui isn't complete.

In general the checkboxes should match the state of the disabled items,
except for "disable open disc" which I don't think people will want to
stay disabled when it gets written.