main: fall back to sensible geometry defaults
This ensures that the windows receive sensible geometry in the case of
not having a saved geometry.json. (i.e. the first run).  The initial
window position in Windows 10 were in the top left in hard to manipulate
locations.

Fetch a copy of the default window geometry state and overwrite with the
saved one (if present).  Also consider having no geometry.json as
another cause for calling the desired geometry functions.