mainwindow: applicationShouldQuit -> instanceShouldQuit
Eventually we're going to have to split the Flow object into two,
and have an AppFlow and a WindowFlow object or so.  One to spawn windows
for multi-window players, and one to control the actual windows.  In an
ideal world, each window would have its own gui thread, but I'm not sure
that's even possible these days.