mainwindow,mpvwidget: add right-click menu
It's a mirror of what is already available in the main menu.  Notable
abscenses: video frame, pan&scan, filters, shaders.

video frame and pan&scan are defintely do-able, but feel like a bit of a
headache and aren't even in the main menu as yet.

shaders needs a slot in the settings window to open the settings window
at the shaders tab.

filters in mpc-hc represent a DirectShow render chain and configuring the
equivalent decoders is not as meaningful to mpv in my experience.  A
roughly equivalent menu would be manipulating the video/audio output
drivers, and as such a "filters" name would be a misnomer; bomi has a
better idea here.

Expanding the context menu and main menu to the robustness of bomi and
mpc-hc needs more work, notably sharing more state between the
mainwindow and mpvwidget.  Along these lines, implementing #112 will
require some sort of signal/option buffering to separate one-time configs
from the mainwindow and permament configs from the settings window.

Fixes #113.