mpvwidget: audio-devices -> audio-device-list
Seems that mpv changed this on me.  Unfortunately I do not have the time
anymore (for about the last 18 months actually) to look at every commit
made to mpv.  Also now that I'm making releases, I'm limited to whatever
the latest release is, so there's another reason not to be ready for the
next mpv release before it arrives.  It's just one of those things.

Fixes audio issue mentioned in #116.