mainwindow,mpvwidget: expose stats.lua
This requires that stats.lua is installed under PREFIX/share/mpc-qt/
scripts, which can be done with the new install set.

Despite my testing, the frame timings page does not want to show for an
unknown reason.  According to the source code, the command `script-
binding stats/display-page-2` will bring up the second stats.lua page.
This does not happen, but it does bring up the first page for a second or
so when the display-stats-toggle is commented out, so at least the
script is doing something.

Fixes #104.