readme: update with qmake + make instructions
Now that our end-users don't have to use QtCreator to compile, there's
no need to nest mpc-qt twice so that ../build-et-cetera doesn't pollute
the main source code directory.  This simplifies the compiling a bit.*

The install set installs the following:
  * PREFIX/bin/mpc-qt
  * PREFIX/share/doc/mpc-qt/ipc.md
  * PREFIX/share/mpc-qt/scripts/stats.lua
  * PREFIX/share/mpc-qt/translations/*.qm
  * PREFIX/share/applications/mpc-qt.desktop
  * PREFIX/share/icons/hicolor/scalable/apps/mpc-qt.svg

Of these, the most notable if stats.lua, because the program expects it
to exist at the above location.

* If you still want to compile with QtCreator, you need to remove the
mpc-qt executable made by qmake+make or it won't go looking for the one
under ../build-et-cetera.