settingswindow,helpers,[others]: implement basic styling
Add the white icon set to resources and create a IconThemer class to
automagically change icons.  Create a section in the settings window to
change application colors.  The palette section needs a bit of a polish,
but this patchset is getting a bit too large for my liking to just put
off the commit indefintely until it is 100% fixed.

Fixes #65.