paletteeditor,settingswindow: add to project
Use a robust palette editor to edit the entire palette rather than a
shortlist of colors.  We were only using the colors needed to construct
a pallete object and generating the rest.  But QPalette::setColor and
QPalette:color support more, so use that instead.  This lets us set the
highlight and tooltip colors, which means that highlight bar no longer
turns into a deep blue no matter what.