main: fix taking screenshots
Wind the screenshot extension into the tempFile earlier.  Logic error in
mainwindow_takeImage meant that the name of the extension of the tempfile
was added to its filename variable after the file was written.  So later
on when file is to be moved to final location, we were trying to move a
file that didn't exist.

Fixes #120.