helpers: match order of MouseState::MousePress enum
Correctly map the text lists and whell mapping to the enum.  If you've
changed the keymapping, you'll have to change them back because the code
is now storing them correctly.

Two root inversions here:
 * pressToText had the wrong order, and therefore was displayed
incorrectly in the settings dialog. (see line 235 of the action editor
module.)  So selecting Down would ask for a up event, and selecting Up
would ask for a down event.
 * fromWheelEvent scroll detect was inverted

This meant that when the default was correctly assigned in source, it
would appear inverted fin the settings dialog and in actual behavour.
But when it was reassigned from the settings window, it was behaving
correctly simply because it went through the inversion twice.  Using
correct behavior fixes this.

Fixes #118.