main,playlistwindow: keep playlists visible after min/restore
Keep the undocked playlist window visible after the application goes
through a minimize->restore change.  The visibilityChanged signal also
triggers when the mainwindow is hidden from minimizing the main app,
which causes the app to forget if the playlist was visible or not when
it's undocked.  The toggleViewAction on the other hand only triggers
through user interaction, so use that instead.