favorites: add to project
What a task.
- add a new favorites window
- write a display list widget for track info
- augment TrackInfo to store (title)text and duration
- put TrackInfo vlist conversion into Helpers module
- add a few signals/slots to manager, mainwindow, flow
- wire up signals

The favorites window has no need for buttons to invoke rename or move up
or move down.  Items can be moved with internal drag/drop, and double
clicking invokes the rename action.  All that might be missing is
intercepting the delete key to remove highlighted items, but this should
be okay for now and I'm not going to let a tiny thing like that stop
progress.