mainwindow: more-correct window frame calculation
Take into account the window frame when calculating desired geometry.
There are two factors at play here:
 - the size of the window frame itself (sizing)
 - the offset of the client area inside the window frame (centering)
The size of the window frame has to be subtracted from the available
desktop area, and the client area offset has to be added to the desired
geometry.

Window frames on windows 10 are transparent except for 1 pixel adjoining
the painted window area, so don't be suprised if you can see around the
edges to the desktop beneath.

Tested on Arch Linux and Windows 10.  Fixes #138.