mpvwidget,[many]: split control code from ui code
Pave the way for changing widget type at runtime and therefore any
future vulkan backend.  Split MpvWidget into Mpv(GlCb/VulkanCb/
Embed)Widget(Interface) and MpvObject.

mpvwidget.cpp now has the following structure:
* interface -> MpvObject
* view -> Mpv(GL/Vulkan/Embed)Widget
* communicator -> MpvController
* other misc helper classes
Which is almost like the usual MVC.

MainWindow and the like do not yet understand the idea that that the
underlying widget could change at any time, we just ask for glcb and be
done with it.  MpvEmbedWidget and MpvVulkanWidget are unimplemented and
lazy typedefs to MpvGlCbWidget.  There are probably some other functions
that could do with a rewrite to better fit this refactor.

Fixes #132.