docs: add coding standards
Describe the way this application has been written.

The majority of the document is simply elaborating on the standard
convention you will encounter in most c++ projects.  Perhaps the most
interesting section to the casual reader would be the links at the top.

This has been sitting in a folder for shy of two years.  Going forward
it might be nice for other developers to know the gist of where the
headspace is.  Feel free to ignore it like I do.